move out some tests
