+++ rust/src/combinators/chain/ffi.rs
@@ -103,7 +103,7 @@ pub extern "C" fn opendp_combinators__make_chain_pm(
 
 #[cfg(test)]
 mod tests {
-    use crate::combinators::tests::{make_test_measurement, make_test_transformation};
+    use crate::combinators::test::{make_test_measurement, make_test_transformation};
     use crate::core::{self, Function};
     use crate::error::Fallible;
     use crate::ffi::any::{AnyObject, Downcast};
